backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
pydantic[email]>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.3.9
anthropic>=0.18.0
aiofiles>=23.2.1
pypdf>=3.17.0
python-docx>=1.1.0
fastapi-cors>=0.0.6
